Second half of first commit

+++ SubCodes/UnKronPolicyIndexes/UnKronPolicyIndexes1_z_e.m
@@ -10,15 +10,15 @@ function Policy=UnKronPolicyIndexes1_z_e(PolicyKron, n_daprime1, n_a, n_z, n_e, 
 
 l_daprime1=length(n_daprime1);
 
-divisors=cumprod([1,n_daprime1(1:end-1)])';   % [l_daprime1,1]
+divisors=cumprod([1,double(n_daprime1(1:end-1))])';   % [l_daprime1,1]
 
 if vfoptions.gridinterplayer==1
-    Policy=[mod(floor((PolicyKron(1,:)-1)./divisors),n_daprime1(:))+1;
+    Policy=[mod(floor(double(PolicyKron(1,:)-1)./divisors),n_daprime1(:))+1;
             PolicyKron(2,:);
             PolicyKron(3,:)];
     Policy=reshape(Policy,[l_daprime1+2,n_a,n_z,n_e]);
 else
-    Policy=mod(floor((PolicyKron-1)./divisors),n_daprime1(:))+1;
+    Policy=mod(floor(double(PolicyKron-1)./divisors),n_daprime1(:))+1;
     Policy=reshape(Policy,[l_daprime1,n_a,n_z,n_e]);
 end
